@@ -3,7 +3,7 @@ description: HarmonyHUB is your ONE source for information related to the Harmon
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.6.0+19
+version: 1.6.0+22
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
@@ -11,13 +11,13 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.0
   font_awesome_flutter: ^8.4.0
   http: ^0.12.0+4
   intl: ">=0.14.0"
   flutter_spinkit: ^4.1.2+1
   shared_preferences: ^0.4.3
-  webview_flutter: ^0.3.22+1
+  webview_flutter: ^2.0.4
   rflutter_alert: ^1.0.3
   firebase_core: ^0.7.0
   firebase_auth: ^0.20.0+1
@@ -34,6 +34,9 @@ dependencies:
   flutter_datetime_picker: ^1.4.0
   flutter_linkify: ^3.1.3
   flutter_native_timezone: ^1.0.4
+  flutter_html:
+
+
 
 
 dev_dependencies:
