@@ -4,6 +4,7 @@ import 'dart:io';
 import 'package:HarmonyHub/components/video_list_item.dart';
 import 'package:HarmonyHub/models/video_sources.dart';
 import 'package:HarmonyHub/screens/info_screen.dart';
+import 'package:HarmonyHub/screens/video_filter_screen.dart';
 import 'package:HarmonyHub/services/video_source_service.dart';
 import 'package:HarmonyHub/youtube_api/youtube_api.dart';
 import 'package:flutter/material.dart';
@@ -21,30 +22,22 @@ class VideosFeedScreen extends StatefulWidget {
   _VideosFeedScreenState createState() => _VideosFeedScreenState();
   static final GlobalKey<_VideosFeedScreenState> videoFeedScreenKey = GlobalKey<_VideosFeedScreenState>();
   void refreshVideos() {
-    videoFeedScreenKey.currentState.callYoutubeAPI();
+    videoFeedScreenKey.currentState.getVideos();
   }
 }
 
 class _VideosFeedScreenState extends State<VideosFeedScreen> {
   bool dataLoading = false;
-  List<Widget> youtubeItems = new List<Widget>();
+  List<Widget> youtubeItems = [];
   YoutubeAPI ytApi = new YoutubeAPI(Secretes.youtubeAccessKey);
-  List<YT_API> videoFeed = new List<YT_API>();
+  List<YT_API> videoFeed = [];
   VideoSourceService vSourceService = new VideoSourceService();
 
-  Future<void> callYoutubeAPI() async {
-    List<VideoSource> vSources = await vSourceService.getVideoSources();
-    if (vSources.length > 0) {
-      String channelId = vSources[0].channelId;
-      videoFeed = await ytApi.channel(channelId);
-      refreshData();
-    }
-  }
-
-  Future<void> refreshData() async {
-    List<Widget> videos = new List<Widget>();
-    for (int i = 0; i < videoFeed.length; i++) {
-      YT_API feed = videoFeed[i];
+  Future<void> getVideos() async {
+    List<Widget> videos = [];
+    List<VideoSource> vSource = await vSourceService.getVideoSources();
+    for (int i = 0; i < vSource.length; i++) {
+      VideoSource feed = vSource[i];
       String title = utf8.decode(utf8.encode(feed.title));
       String description = utf8.decode(utf8.encode(feed.description));
       //print(feed.url);
@@ -53,13 +46,13 @@ class _VideosFeedScreenState extends State<VideosFeedScreen> {
         height: description.length < 75 ? 80 : 135,
         title: title,
         text: description,
-        leading: feed.thumbnail["default"]["url"] == null
+        leading: feed.thumbnailURL == null
             ? Icon(
                 FontAwesomeIcons.youtube,
                 size: 40,
               )
             : Image.network(
-                feed.thumbnail["default"]["url"],
+                feed.thumbnailURL,
               ),
         moreDetails: true,
         openMoreDetails: () async {
@@ -69,12 +62,12 @@ class _VideosFeedScreenState extends State<VideosFeedScreen> {
               MaterialPageRoute(
                 builder: (context) => InformationScreen(
                   title: feed.title,
-                  url: "https://www.youtube.com/watch?v=${feed.id}",
+                  url: feed.videoUrl,
                 ),
               ),
             );
           } else {
-            String url = "https://www.youtube.com/watch?v=${feed.id}";
+            String url = feed.videoUrl;
             await canLaunch(url) ? await launch(url) : print('Could not launch $url');
           }
         },
@@ -91,7 +84,7 @@ class _VideosFeedScreenState extends State<VideosFeedScreen> {
   @override
   void initState() {
     super.initState();
-    //callYoutubeAPI();
+    getVideos();
   }
 
   @override
@@ -106,13 +99,28 @@ class _VideosFeedScreenState extends State<VideosFeedScreen> {
             'assets/app_icon.png',
           ),
         ),
+        actions: [
+          new IconButton(
+            icon: Icon(FontAwesomeIcons.filter),
+            onPressed: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => VideoFilterScreen(
+                    refreshVideos: getVideos,
+                  ),
+                ),
+              );
+            },
+          ),
+        ],
         iconTheme: IconThemeData(
           color: Colors.white, //change your color here
         ),
       ),
       body: Container(
         child: RefreshIndicator(
-          onRefresh: callYoutubeAPI,
+          onRefresh: getVideos,
           child: ListView(
             padding: const EdgeInsets.all(5),
             children: <Widget>[
