@@ -5,13 +5,27 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_core/firebase_core.dart';
 
 class VideoSourceService {
-  final String videoSourcesTable = "video_sources";
+  final String videoSourcesTable = "videos_metadata";
   Future<List<VideoSource>> getVideoSources() async {
     await Firebase.initializeApp();
     FirebaseAuthService fas = new FirebaseAuthService();
     String userId = await fas.getUserId();
     String projectId = await Global.getProjectId();
     if (userId != null) {
+      List<String> filters = await Global.getUserFavList(Global.favoriteVideoHandlesKey);
+      bool includeSearchItems = false;
+      List<String> selectedProviders = [];
+      for (String filter in filters) {
+        List<String> parts = filter.split(":##:");
+        if (parts.length > 1) {
+          if (parts[1] == 'search') {
+            includeSearchItems = true;
+          }
+        }
+        if (parts.length > 2) {
+          selectedProviders.add(parts[2]);
+        }
+      }
       Global.videoSources.clear();
       QuerySnapshot videoSources = await FirebaseFirestore.instance
           .collection(videoSourcesTable)
@@ -20,17 +34,25 @@ class VideoSourceService {
             isEqualTo: projectId,
           )
           .get();
-      List<String> docIds = new List<String>();
+      List<String> docIds = [];
       for (int i = 0; i < videoSources.docs.length; i++) {
         var element = videoSources.docs[i];
-        docIds.add(element.id);
-        VideoSource vSource = VideoSource(
-          channelId: element['channel_id'],
-          source: element['source'],
-          url: element['url'],
-          projectId: element['project_id'],
-        );
-        Global.videoSources.add(vSource);
+        String provider = element['video_provider'];
+        if (selectedProviders.contains(provider) || includeSearchItems) {
+          docIds.add(element.id);
+          VideoSource vSource = VideoSource(
+            docId: element.id,
+            title: element['title'],
+            description: element['description'],
+            videoDatetime: element['video_datetime'].toDate(),
+            projectId: element['project_id'],
+            videoProvider: element['video_provider'],
+            videoSource: element['video_source'],
+            videoUrl: element['video_url'],
+            thumbnailURL: element['video_thumbnail_url'],
+          );
+          Global.videoSources.add(vSource);
+        }
       }
       return Global.videoSources;
     }
